feat: add winner videos section and update support details in GulfTicketAppSection
